requirements.txt
# ── Attestix Protocol ────────────────────────────────────────────────────────
attestix==0.3.0

# ── LLM Inference ────────────────────────────────────────────────────────────
groq==0.37.1
openai==2.36.0

# ── Agent Frameworks ─────────────────────────────────────────────────────────
langchain==1.2.18
langchain-core==1.3.3
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-text-splitters==1.1.2
langgraph==1.1.10
crewai==1.14.4
semantic-kernel==1.41.3

# ── Blockchain (optional — only needed for Base Sepolia anchoring) ────────────
web3==7.16.0
eth-account==0.13.7

# ── Environment ──────────────────────────────────────────────────────────────
python-dotenv==1.2.2
